Fixed a spatial query bug

+++ RealVirtualIntegration/WebServer/src/com/cyberlightning/webserver/entities/RowEntry.java
@@ -12,7 +12,7 @@ public class RowEntry implements Serializable {
 	public String latestEvent;
 	public String entityUUID;
 	public String contextUUID;
-	public float[] location;
+	public double[] location;
 	
 
 	public RowEntry () {
@@ -23,7 +23,7 @@ public class RowEntry implements Serializable {
 	}
 	
 	public void setLocation(Float _lat, Float _lon) {
-		this.location = new float[2];
+		this.location = new double[2];
 		this.location[0] = _lat;
 		this.location[1] = _lon;
 	}
